+++ src/main/scanner/models/rectangle.hpp
@@ -5,11 +5,11 @@
 
 class rectangle {
     public:
-        Eigen::Vector2f UL, BR;
+        Eigen::Vector2d UL, BR;
         float w, h;
 
-        rectangle(const Eigen::Vector2f& ul, const Eigen::Vector2f& br);
-        rectangle translate(const Eigen::Vector2f& r0);      
+        rectangle(const Eigen::Vector2d& ul, const Eigen::Vector2d& br);
+        rectangle translate(const Eigen::Vector2d& r0);      
 };
 
 #endif
